improve softmax/rmsnorm/layernorm performance
