+++ python/examples/__init__.py
@@ -1,0 +1,7 @@
+"""Reusable RocDSL Python examples and kernel builders.
+
+This directory is a Python package so tests can import kernel builders via:
+`from examples.layernorm_kernel import build_layernorm_module`.
+"""
+
+
