+++ python/examples/reduce.py
@@ -1,0 +1,218 @@
+"""Shared reduction helpers for RocDSL example kernels.
+
+These helpers build MLIR ops (rocir/gpu/scf/vector/etc). They are extracted from
+softmax/layernorm/rmsnorm kernels to de-duplicate code without changing codegen.
+"""
+
+def unwrap(v):
+    if hasattr(v, "value"):
+        return v.value
+    if hasattr(v, "_value"):
+        return v._value
+    if hasattr(v, "result"):
+        return v.result
+    return v
+
+
+def reduce_vec_max(vec_val, *, VEC_WIDTH, compute_type, vector):
+    if VEC_WIDTH == 1:
+        return vector.extract(vec_val, static_position=[0], dynamic_position=[])
+    # Avoid fastmath on bf16 max reduction; some backends can fail to select.
+    return vector.reduction(compute_type, "maxnumf", unwrap(vec_val))
+
+
+def reduce_vec_sum(vec_val, *, VEC_WIDTH, compute_type, vector, fm_fast):
+    if VEC_WIDTH == 1:
+        return vector.extract(vec_val, static_position=[0], dynamic_position=[])
+    return vector.reduction(compute_type, "add", unwrap(vec_val), fastmath=fm_fast)
+
+
+def make_block_reduce(*, tid, BLOCK_SIZE, compute_type, arith, gpu, rocir, s_red_tv, T, ir, c_zero, c_neg_inf, c_zero_idx, fm_fast):
+    """Return a `block_reduce(val, reduce_op_name)` function (softmax-style)."""
+
+    def block_reduce(val, reduce_op_name):
+        # AMD wavefront size is 64 on gfx9+/gfx10+/gfx11.
+        WARP_SIZE = 64
+        NUM_WAVES = (BLOCK_SIZE + WARP_SIZE - 1) // WARP_SIZE  # python int
+        # Use Rocir layout algebra to compute LDS indices for the reduction scratch.
+        c_num_waves = rocir.const_index(NUM_WAVES)
+        c1 = rocir.const_index(1)
+        shape_red = rocir.make_shape(unwrap(c_num_waves))
+        stride_red = rocir.make_stride(unwrap(c1))
+        layout_red = rocir.make_layout(shape_red, stride_red)
+
+        tid_i32 = rocir.arith.IndexCastOp(T.i32(), unwrap(tid)).result
+        c_warp_i32 = arith.constant(WARP_SIZE, type=T.i32()).value
+        lane_i32 = rocir.arith.RemUIOp(unwrap(tid_i32), unwrap(c_warp_i32)).result
+        wave_i32 = rocir.arith.DivUIOp(unwrap(tid_i32), unwrap(c_warp_i32)).result
+
+        width_i32 = arith.constant(WARP_SIZE, type=T.i32()).value
+        w = unwrap(val)
+
+        # Intra-wave reduction via xor shuffle
+        for sh in [32, 16, 8, 4, 2, 1]:
+            off = arith.constant(sh, type=T.i32()).value
+            peer = gpu.ShuffleOp(unwrap(w), unwrap(off), unwrap(width_i32), mode="xor").shuffleResult
+            if reduce_op_name == "max":
+                w = rocir.arith.MaximumFOp(unwrap(w), unwrap(peer)).result
+            else:
+                w = rocir.arith.AddFOp(unwrap(w), unwrap(peer), fastmath=fm_fast).result
+
+        # lane0 writes per-wave partial into LDS s_red[wave_id]
+        is_lane0 = rocir.arith.CmpIOp(
+            rocir.arith.CmpIPredicate.eq,
+            unwrap(lane_i32),
+            unwrap(arith.constant(0, type=T.i32()).value),
+        ).result
+        if_lane0 = rocir.scf_ext.IfOp(unwrap(is_lane0))
+        with ir.InsertionPoint(if_lane0.then_block):
+            wave_idx = rocir.arith.IndexCastOp(T.index(), unwrap(wave_i32)).result
+            red_idx = rocir.crd2idx(rocir.make_coord(unwrap(wave_idx)), layout_red)
+            s_red_tv[unwrap(red_idx)] = unwrap(w)
+            rocir.scf_ext.yield_([])
+        gpu.barrier()
+
+        # wave0 reduces NUM_WAVES partials (still using shuffle)
+        is_wave0 = rocir.arith.CmpIOp(
+            rocir.arith.CmpIPredicate.eq,
+            unwrap(wave_i32),
+            unwrap(arith.constant(0, type=T.i32()).value),
+        ).result
+        if_wave0 = rocir.scf_ext.IfOp(unwrap(is_wave0), [compute_type], hasElse=True)
+        with ir.InsertionPoint(if_wave0.then_block):
+            in_range = rocir.arith.CmpIOp(
+                rocir.arith.CmpIPredicate.ult,
+                unwrap(lane_i32),
+                unwrap(arith.constant(NUM_WAVES, type=T.i32()).value),
+            ).result
+            if_in = rocir.scf_ext.IfOp(unwrap(in_range), [compute_type], hasElse=True)
+            with ir.InsertionPoint(if_in.then_block):
+                lane_idx = rocir.arith.IndexCastOp(T.index(), unwrap(lane_i32)).result
+                red_idx = rocir.crd2idx(rocir.make_coord(unwrap(lane_idx)), layout_red)
+                v = s_red_tv[unwrap(red_idx)]
+                rocir.scf_ext.yield_([unwrap(v)])
+            with ir.InsertionPoint(if_in.else_block):
+                neutral = c_neg_inf if reduce_op_name == "max" else c_zero
+                rocir.scf_ext.yield_([unwrap(neutral)])
+
+            ww = if_in.results[0]
+            for sh in [32, 16, 8, 4, 2, 1]:
+                off = arith.constant(sh, type=T.i32()).value
+                peer = gpu.ShuffleOp(unwrap(ww), unwrap(off), unwrap(width_i32), mode="xor").shuffleResult
+                if reduce_op_name == "max":
+                    ww = rocir.arith.MaximumFOp(unwrap(ww), unwrap(peer)).result
+                else:
+                    ww = rocir.arith.AddFOp(unwrap(ww), unwrap(peer), fastmath=fm_fast).result
+
+            # lane0 writes final to s_red[0]
+            is_lane0_2 = rocir.arith.CmpIOp(
+                rocir.arith.CmpIPredicate.eq,
+                unwrap(lane_i32),
+                unwrap(arith.constant(0, type=T.i32()).value),
+            ).result
+            if_lane0_2 = rocir.scf_ext.IfOp(unwrap(is_lane0_2))
+            with ir.InsertionPoint(if_lane0_2.then_block):
+                red_idx0 = rocir.crd2idx(rocir.make_coord(unwrap(c_zero_idx)), layout_red)
+                s_red_tv[unwrap(red_idx0)] = unwrap(ww)
+                rocir.scf_ext.yield_([])
+
+            rocir.scf_ext.yield_([unwrap(ww)])
+        with ir.InsertionPoint(if_wave0.else_block):
+            red_idx0 = rocir.crd2idx(rocir.make_coord(unwrap(c_zero_idx)), layout_red)
+            keep = s_red_tv[unwrap(red_idx0)]
+            rocir.scf_ext.yield_([unwrap(keep)])
+        gpu.barrier()
+
+        red_idx0 = rocir.crd2idx(rocir.make_coord(unwrap(c_zero_idx)), layout_red)
+        return s_red_tv[unwrap(red_idx0)]
+
+    return block_reduce
+
+
+def make_block_reduce_add(*, tid, fm_fast, WARP_SIZE, RED_SLOTS, gpu, arith, arith_ops, rocir, T, ir, zero_idx, scratch_tv_shape_stride=(None, None)):
+    """Return a `block_reduce_add(val_f32, scratch_memref)` function (norm-style)."""
+    shape_unused, stride_unused = scratch_tv_shape_stride
+    _ = shape_unused
+    _ = stride_unused
+
+    def block_reduce_add(val_f32, scratch_memref):
+        scratch_tv = rocir.make_tensor(scratch_memref, shape=(RED_SLOTS,), strides=(1,))
+        tid_i32 = arith_ops.IndexCastOp(T.i32(), tid.value).result
+        c_warp_i32 = arith.constant(T.i32(), WARP_SIZE)
+        lane_i32 = arith_ops.RemUIOp(unwrap(tid_i32), unwrap(c_warp_i32)).result
+        wave_i32 = arith_ops.DivUIOp(unwrap(tid_i32), unwrap(c_warp_i32)).result
+        width_i32 = arith.constant(T.i32(), WARP_SIZE)
+        # Use Rocir layout algebra to compute LDS indices for the reduction scratch.
+        c_num_waves = rocir.const_index(RED_SLOTS)
+        c1 = rocir.const_index(1)
+        shape_red = rocir.make_shape(unwrap(c_num_waves))
+        stride_red = rocir.make_stride(unwrap(c1))
+        layout_red = rocir.make_layout(shape_red, stride_red)
+
+        w = unwrap(val_f32)
+        for sh in [32, 16, 8, 4, 2, 1]:
+            off = arith.constant(T.i32(), sh)
+            peer = gpu.ShuffleOp(unwrap(w), unwrap(off), unwrap(width_i32), mode="xor").shuffleResult
+            w = arith_ops.AddFOp(unwrap(w), unwrap(peer), fastmath=fm_fast).result
+
+        is_lane0 = arith_ops.CmpIOp(
+            arith_ops.CmpIPredicate.eq,
+            unwrap(lane_i32),
+            unwrap(arith.constant(T.i32(), 0)),
+        ).result
+        if_lane0 = rocir.scf_ext.IfOp(unwrap(is_lane0))
+        with ir.InsertionPoint(if_lane0.then_block):
+            wave_idx = arith_ops.IndexCastOp(T.index(), unwrap(wave_i32)).result
+            red_idx = rocir.crd2idx(rocir.make_coord(unwrap(wave_idx)), layout_red)
+            scratch_tv[unwrap(red_idx)] = unwrap(w)
+            rocir.scf_ext.yield_([])
+        gpu.barrier()
+
+        NUM_WAVES = RED_SLOTS
+        is_wave0 = arith_ops.CmpIOp(
+            arith_ops.CmpIPredicate.eq,
+            unwrap(wave_i32),
+            unwrap(arith.constant(T.i32(), 0)),
+        ).result
+        # Only wave0 does final reduction and writes scratch[0].
+        if_wave0 = rocir.scf_ext.IfOp(unwrap(is_wave0))
+        with ir.InsertionPoint(if_wave0.then_block):
+            in_range = arith_ops.CmpIOp(
+                arith_ops.CmpIPredicate.ult,
+                unwrap(lane_i32),
+                unwrap(arith.constant(T.i32(), NUM_WAVES)),
+            ).result
+            if_in = rocir.scf_ext.IfOp(unwrap(in_range), [T.f32()], hasElse=True)
+            with ir.InsertionPoint(if_in.then_block):
+                lane_idx = arith_ops.IndexCastOp(T.index(), unwrap(lane_i32)).result
+                red_idx = rocir.crd2idx(rocir.make_coord(unwrap(lane_idx)), layout_red)
+                v = scratch_tv[unwrap(red_idx)]
+                rocir.scf_ext.yield_([unwrap(v)])
+            with ir.InsertionPoint(if_in.else_block):
+                rocir.scf_ext.yield_([unwrap(arith.constant(T.f32(), 0.0).value)])
+
+            ww = if_in.results[0]
+            for sh in [32, 16, 8, 4, 2, 1]:
+                off = arith.constant(T.i32(), sh)
+                peer = gpu.ShuffleOp(unwrap(ww), unwrap(off), unwrap(width_i32), mode="xor").shuffleResult
+                ww = arith_ops.AddFOp(unwrap(ww), unwrap(peer), fastmath=fm_fast).result
+
+            is_lane0_2 = arith_ops.CmpIOp(
+                arith_ops.CmpIPredicate.eq,
+                unwrap(lane_i32),
+                unwrap(arith.constant(T.i32(), 0)),
+            ).result
+            if_lane0_2 = rocir.scf_ext.IfOp(unwrap(is_lane0_2))
+            with ir.InsertionPoint(if_lane0_2.then_block):
+                red_idx0 = rocir.crd2idx(rocir.make_coord(unwrap(zero_idx)), layout_red)
+                scratch_tv[unwrap(red_idx0)] = unwrap(ww)
+                rocir.scf_ext.yield_([])
+            rocir.scf_ext.yield_([])
+
+        gpu.barrier()
+        red_idx0 = rocir.crd2idx(rocir.make_coord(unwrap(zero_idx)), layout_red)
+        return scratch_tv[unwrap(red_idx0)]
+
+    return block_reduce_add
+
+
